Revert "refactor(theme): move qconfig patching to its own file"
This reverts commit 69c03f7eb770346d085ac178e2cb921a932bc5e5.
